requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.3
streamlit==1.39.0
plotly==5.24.1
